services/wallet/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.6.1
grpcio>=1.68.0
protobuf>=5.29.0
cryptography>=44.0.0
sqlalchemy>=2.0.36
greenlet>=3.1.1
asyncpg>=0.30.0
python-jose[cryptography]>=3.3.0
pyotp>=2.9.0
httpx>=0.27.0
embit>=0.7.0
